+++ demo/src/main/java/org/vaadin/stefan/ui/layouts/AbstractLayout.java
@@ -44,7 +44,7 @@ import java.util.Locale;
 @CssImport("./app-layout-styles.css")
 @SuppressWarnings("rawtypes")
 public abstract class AbstractLayout extends AppLayout implements AfterNavigationObserver {
-    public static final String ADDON_VERSION = "6.1.1-SNAPSHOT";
+    public static final String ADDON_VERSION = "6.2.0";
     private static final long serialVersionUID = -7479612679602267287L;
 
     @SuppressWarnings("unchecked")